Reorganise and tweak CLI, add deep merge to spec
